builing controller for Favorite songs and fixing

+++ backend/Spotify/src/main/java/com/eoi/spotify/service/UserService.java
@@ -5,11 +5,11 @@ import com.eoi.spotify.entity.User;
 import com.eoi.spotify.projection.UserProjection;
 import com.eoi.spotify.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
 public class UserService {
@@ -23,17 +23,38 @@ public class UserService {
         return ur.findAllUsersProjected();
     }
 
-    public Optional<User> getUserById(int id) {
-        return ur.findById(id);
+    public User getUserById(int id) {
+        return ur.findById(id)
+                .orElseThrow(() ->
+                        new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found", null));
     }
     public User insertUser(User u) {
         u.setId(0);
         return ur.save(u);
     }
-    public User updateUser(User u) {
-        return ur.save(u);
+    public User updateUser(User u, int id) {
+        User existingUser = ur.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User Not found", null));
+
+        if (u.getImage() != null) {
+            existingUser.setImage(u.getImage());
+        }
+        if (u.getPassword() != null) {
+            existingUser.setPassword(u.getPassword());
+        }
+        if (u.getMail() != null) {
+            existingUser.setMail(u.getMail());
+        }
+        if (u.getName() != null) {
+            existingUser.setName(u.getName());
+        }
+
+        return ur.save(existingUser);
     }
-    public void deleteUser(int id) {
+    public void deleteUserById(int id) {
+        if(!ur.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND ,"User not found");
+        }
         ur.deleteById(id);
     }
